refactor: trip log comment, like service를 분리 + 삭제된 댓글에 대댓글 못 달도록 변경
